requirements.txt
asgiref>=3.6.0
certifi==2022.9.24
charset-normalizer==2.1.1
django>=4.2
django-webhook-consume@ git+https://github.com/FlorianRupp/django-webhook-consume.git@eb895b00c59286a79c921ec91ae9a30c4dc2b816
gunicorn==20.1.0
importlib-metadata==4.6.4
isodate==0.6.1
Markdown==3.3.6
more-itertools==8.10.0
powerline-status==2.7
Pygments==2.11.2
pyparsing==3.0.9
requests>=2.28.1
six==1.16.0
SPARQLWrapper==2.0.0
sqlparse==0.4.3
urllib3==1.26.12
zipp==1.0.0
channels>=2.4.0
elasticsearch>=7.1.0
libsass==0.20.0
django-compressor>=4.5.1
django-crispy-forms>=1.13.1
django-sass-processor==0.8
django-hcaptcha==0.2.0
django-active-link==0.1.8
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
cffi==1.15.1
constantly==15.1.0
crispy-bootstrap5==0.7
cryptography==38.0.1
daphne==3.0.2
Deprecated==1.2.13
discord==2.0.0
discord.py==2.0.1
django-environ==0.9.0
django-hCaptcha==0.2.0
django-ranged-response==0.2.0
django-redis==5.2.0
django-simple-captcha==0.5.17
elastic-transport==8.4.0
elasticsearch==8.4.3
environ==1.0
frozenlist==1.3.1
hyperlink==21.0.0
incremental==21.3.0
multidict==6.0.2
packaging==21.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
PyYAML==6.0
rdflib==6.2.0
requests-file==2.1.0
requests-oauthlib==1.3.1
rjsmin>=1.2.2
Scrapy==2.12.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.6
SPARQLWrapper==2.0.0
sqlparse==0.4.3
tldextract==5.1.3
redis>=4.3.4
regex==2022.9.13
toml==0.10.2
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
wrapt==1.14.1
yarl==1.8.1
zope.interface==5.5.0
pandas==1.5.0
rdflib==6.2.0
django-cookiebanner>=0.2.8
djangorestframework>=3.14.0
django-debug-toolbar>=4.2
django-extensions>=3.2.3
django-webpack-loader>=2.0
drf-yasg>=1.21.7
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
pysolr>=3.0.0
